Refactor user table variable userId and empNum convert long into int

+++ backend/daoumarket/src/main/java/com/daoumarket/dao/IFavoriteDao.java
@@ -7,5 +7,5 @@ import com.daoumarket.dto.Favorite;
 public interface IFavoriteDao {
 	public int insertFavorite(Favorite favorite);
 	public int deleteFavorite(long favoriteId);
-	public List<Favorite> getFavoriteList(long userId);
+	public List<Favorite> getFavoriteList(int userId);
 }
